+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### added
 * support for positional parameters with default values
 * support for `typing.TypeVar`
+* added `MultiDispatch.candidates()` method
 ### changed
 * `float` is now interpreted as `Union[float, int]`
 * `complex` is now interpreted as `Union[float, int, complex]`